# Integrate PrecomputedText API, add Performance Sample.