--- a/samples/mvvm/build.gradle
+++ b/samples/mvvm/build.gradle
@@ -33,7 +33,6 @@ dependencies {
     implementation libraries.kotlinStdlib
 
     implementation supportLibraries.appCompat
-    implementation supportLibraries.recyclerView
     implementation supportLibraries.constraintLayout
 
     implementation libraries.rxJava