--- a/samples/performance/src/main/java/com/lyft/domic/samples/performance/domic/DomicView.kt
+++ b/samples/performance/src/main/java/com/lyft/domic/samples/performance/domic/DomicView.kt
@@ -20,7 +20,4 @@ interface DomicView {
     val button: Button
     val checkBox: CompoundButton
     val radioButton: CompoundButton
-
-    val editText1: EditText
-    val editText2: EditText
 }