--- a/samples/performance/src/main/java/com/lyft/domic/samples/performance/regular/RegularView.kt
+++ b/samples/performance/src/main/java/com/lyft/domic/samples/performance/regular/RegularView.kt
@@ -1,0 +1,18 @@
+package com.lyft.domic.samples.performance.regular
+
+interface RegularView {
+    fun setCounter0Text(text: CharSequence)
+    fun setCounter1Text(text: CharSequence)
+    fun setCounter2Text(text: CharSequence)
+    fun setCounter3Text(text: CharSequence)
+    fun setCounter4Text(text: CharSequence)
+    fun setCounter5Text(text: CharSequence)
+    fun setCounter6Text(text: CharSequence)
+    fun setCounter7Text(text: CharSequence)
+    fun setCounter8Text(text: CharSequence)
+    fun setCounter9Text(text: CharSequence)
+
+    fun setButtonEnabled(enabled: Boolean)
+    fun setCheckBoxChecked(checked: Boolean)
+    fun setRadioButtonChecked(checked: Boolean)
+}