--- a/settings.gradle
+++ b/settings.gradle
@@ -5,4 +5,5 @@ include ':domic:util'
 
 include ':samples:mvvm'
 include ':samples:mvp'
+include ':samples:performance'
 include ':samples:shared'