… samples/performance/src/main/java/com/lyft/domic/samples/performance/regular/RegularPresenter.kt
… samples/performance/src/main/res/layout/activity_performance.xml
… samples/performance/src/main/res/mipmap-hdpi/ic_launcher.png
… samples/performance/src/main/res/mipmap-hdpi/ic_launcher_round.png
… samples/performance/src/main/res/mipmap-mdpi/ic_launcher.png
… samples/performance/src/main/res/mipmap-mdpi/ic_launcher_round.png
… samples/performance/src/main/res/mipmap-xhdpi/ic_launcher.png
… samples/performance/src/main/res/mipmap-xhdpi/ic_launcher_round.png
… samples/performance/src/main/res/mipmap-xxhdpi/ic_launcher.png
… samples/performance/src/main/res/mipmap-xxhdpi/ic_launcher_round.png
… samples/performance/src/main/res/mipmap-xxxhdpi/ic_launcher.png
… samples/performance/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png