Closes #8.

***Disclaimer***: Performance is not the reason of Domic's existence, however because of Domic's design we're able to achieve ridiculous things and I honestly believe that Domic will be 120 fps ready once we hit v1 (but don't quote me on that lol).

[PrecomputedText is an API](https://developer.android.com/reference/android/text/PrecomputedText) designed to allow pre-computing text layouting data before actual rendering on bg threads thus avoiding these computations on Main and/or rendering thread, which PERFECTLY fits into Domic's design

Note that there is a support version of this API https://developer.android.com/reference/androidx/core/text/PrecomputedTextCompat, but I'm not sure about the support lib relations with Domic yet, so for now we depend only on Framework APIs.

---

PR also adds ridiculous Performance Sample that renders 10 TextView counters and toggles 3 buttons at **insane rates** (each 0.999 ms, 1ms, 2ms, etc).

![compressed gif](https://user-images.githubusercontent.com/967132/42439131-d4c033d6-8316-11e8-92ba-30c4ec34f8b9.gif)

Systrace of this UI without Domic (same threading, only actual rendering switches to Main Thread):

(code in `samples.performance.regular` package)

<img width="1428" alt="screen shot 2018-07-09 at 1 09 36 am" src="https://user-images.githubusercontent.com/967132/42439181-f36cff08-8316-11e8-8f93-06bb392fd6be.png">

Systrace of this UI with Domic and PrecomputedText:
(code in `samples.performance.domic` package)

<img width="1274" alt="screen shot 2018-07-09 at 1 10 59 am" src="https://user-images.githubusercontent.com/967132/42439323-7226fd80-8317-11e8-81b9-4583e340f756.png">

Systrace of this UI with Domic and without PrecomputedText:

<img width="1331" alt="screen shot 2018-07-09 at 1 29 31 am" src="https://user-images.githubusercontent.com/967132/42439367-9769109c-8317-11e8-9819-6ca94f5a1e22.png">

Note that **in this particular example Domic's diffing engine doesn't optimize anything** because each new rendered value is different from the old one, I want to showcase that in the sample some time in the future. 

**The sole reason for such good results is our optimized double-buffer rendering + sync with Chroreographer.**

There is also noticeable difference in gc kick-in frequency between Precomputed vs no Precomputed text systraces, I'll be taking a look later, might have been just runtime decision of ART.

Environment:

- Android Emulator 27.1.12.0 (Nexus 5X, API 28, Google Play Image)
- 3 GHz Intel Xeon W
- 64 GB 2666 MHz DDR4
- Radeon Pro Vega 64 16368 MB
- macOS 10.13.5