@artem-zinnatullin 2018-07-09
>Isn't it possible to reuse some of the Views / ViewModel between the comparision (normal + regular)?

It is possible to a degree, I could also have `List<TextView>` instead of separate TextViews but after all that will complicate code and debugging/tracing quite significantly and last thing I want in Performance Sample is complicated code